feat(App.jsx, assets/projects.js, projects/AcnhPlayer.jsx): add ACNH Music Player project
Add a new project to the website, ACNH Music Player, which is a music player app that queries the ACNH API for the background music corresponding to the user's local time. It updates the next song at the top of the hour and has weather selection. The project uses HTML, CSS, React, and Tailwind. The commit also includes the addition of the project's thumbnail image and a screenshot to the assets folder, as well as the creation of the AcnhPlayer component.